Module nimbus::client::http_client

This is a simple HTTP client that uses viaduct to retrieve experiment data from the server. Currently configured to use Kinto and the old schema, although that would change once we start working on the real Nimbus schema.

In the future we might replace this with a more fully-feature Remote Settings client, such as:

https://github.com/mozilla-services/remote-settings-client Issue: https://github.com/mozilla/application-services/issues/3475

But the simple subset implemented here meets our needs for now.

Structs

Client

Enums

RemoteState

Constants

HEADER_BACKOFF

HEADER_RETRY_AFTER

Functions

parse_experiments