Function nimbus::evaluator::choose_branch

fn choose_branch<'a>(
    slug: &str, 
    branches: &'a [Branch], 
    id: &str
) -> Result<&'a Branch>

Chooses a branch randomly from a set of branches based on the ratios set in the branches

It is important that the input to the sampling algorithm be:

Unique per-user (no one is bucketed alike)

Unique per-experiment (bucketing differs across multiple experiments)

Differs from the input used for sampling the recipe (otherwise only branches that contain the same buckets as the recipe sampling will receive users)

Arguments:

slug the slug associated with the experiment

branches the branches to pick from

id the user id used to pick a branch

Returns:

Returns the slug for the selected branch

Errors:

An error could occur if something goes wrong while sampling the ratios