Function nimbus::evaluator::targeting

fn targeting(
    expression_statement: &str, 
    ctx: &AppContext
) -> Option<EnrollmentStatus>

Checks if the client is targeted by an experiment This api evaluates the JEXL statement retrieved from the server against the application context provided by the client

Arguments

expression_statement: The JEXL statement provided by the server

ctx: The application context provided by the client

If this app can not be targeted, returns an EnrollmentStatus to indicate why. Returns None if we should continue to evaluate the enrollment status.

In practice, if this returns an EnrollmentStatus, it will be either EnrollmentStatus::NotEnrolled, or EnrollmentStatus::Error in the following cases (But not limited to):

The expression_statement is not a valid JEXL statement

The expression_statement expects fields that do not exist in the AppContext definition

The result of evaluating the statement against the context is not a boolean

jexl-rs returned an error