Module nimbus::matcher

This module defines all the information needed to match a user with an experiment. Soon it will also include a match function of some sort that does the matching.

It has two main types, the Matcher retrieved from the server, and the AppContext provided by the consuming client.

Structs

AppContext
The AppContext object represents the parameters and characteristics of the consuming application that we are interested in for targeting purposes. The app_name, app_id, and channel fields are not optional as they are expected to be provided by all consuming applications as they are used in the first pieces of targeting that help to ensure that an experiment is only processed by the correct application.

Matcher