Module nimbus::persistence

Our storage abstraction, currently backed by Rkv.

Re-exports

pub use backend::Readable;

pub use backend::Writer;

Modules

backend

Structs

Database
Database used to access persisted data This an abstraction around an Rkv database An instance on this database is created each time the component is loaded if there is persisted data, the get functions should retrieve it

SingleStore
A wrapper for an Rkv store. Implemented to allow any value which supports serde to be used.

Enums

StoreId
Enumeration of the different stores within our database.

Constants

DB_KEY_DB_VERSION

DB_VERSION