Function nimbus::sampling::bucket_sample

pub(crate) fn bucket_sample<T: Serialize>(
    input: T, 
    start: u32, 
    count: u32, 
    total: u32
) -> Result<bool>

Sample by splitting the input space into a series of buckets, checking if the given input is in a range of buckets

The range to check is defined by a start point and length, and can wrap around the input space. For example, if there are 100 buckets, and we ask to check 50 buckets starting from bucket 70, then buckets 70-99 and 0-19 will be checked

Arguments:

input What will be hashed and matched against the range of the buckets

start the index of the bucket to start checking

count then number of buckets to check

total The total number of buckets to group inputs into

Returns:

Returns true if the hash generated from the input belongs within the range otherwise false

Errors:

Could error in the following cases (but not limited to)

An error occured in the hashing process

an error occured while checking if the hash belongs in the bucket