Function nimbus::sampling::ratio_sample

pub(crate) fn ratio_sample<T: Serialize>(
    input: T, 
    ratios: &[u32]
) -> Result<usize>

Sample over a list of ratios such that, over the input space, each ratio has a number of matches in correct proportion to the other ratios

Arguments:

input: the input used in the sampling process

ratios: The list of ratios associated with each option

Example:

Assuming the ratios: [1, 2, 3, 4] 10% of all inputs will return 0, 20% will return 1 and so on

Returns

Returns an index of the ratio that matched the input

Errors

Could return an error if the input couldn't be hashed