nimbus::stateful::gecko_prefs

Trait GeckoPrefHandler

pub trait GeckoPrefHandler: Send + Sync {
    // Required methods
    fn get_prefs_with_state(
        &self,
    ) -> MapOfFeatureIdToPropertyNameToGeckoPrefState;
    fn set_gecko_prefs_state(&self, new_prefs_state: Vec<GeckoPrefState>);
    fn set_gecko_prefs_original_values(
        &self,
        original_gecko_prefs: Vec<OriginalGeckoPref>,
    );
}

Required Methods

fn get_prefs_with_state(&self) -> MapOfFeatureIdToPropertyNameToGeckoPrefState

Used to obtain the prefs values from Gecko

fn set_gecko_prefs_state(&self, new_prefs_state: Vec<GeckoPrefState>)

Used to set the state for each pref based on enrollments

fn set_gecko_prefs_original_values( &self, original_gecko_prefs: Vec<OriginalGeckoPref>, )

Used to set back to the original state for each pref based on the original Gecko value