nimbus::stateful
Module nimbus_client

Structs

InternalMutableState

NimbusClient

Nimbus is the main struct representing the experiments state It should hold all the information needed to communicate a specific user’s experimentation status

NimbusStringHelper

Constants

DB_KEY_APP_VERSION

DB_KEY_FETCH_ENABLED

DB_KEY_INSTALLATION_DATE

DB_KEY_UPDATE_DATE