Struct nimbus::stateful::persistence::Database

pub struct Database { /* private fields */ }

Database used to access persisted data This an abstraction around an Rkv database An instance on this database is created each time the component is loaded if there is persisted data, the get functions should retrieve it

Implementations

impl Database

pub fn new<P: AsRef<Path>>(path: P) -> Result<Self>

Main constructor for a database Initiates the Rkv database to be used to retrieve persisted data

Arguments

path: A path to the persisted data, this is provided by the consuming application

pub fn get_store(&self, store_id: StoreId) -> &SingleStore

Gets a Store object, which used with the writer returned by self.write() to update the database in a transaction.

pub fn open_rkv<P: AsRef<Path>>(path: P) -> Result<Rkv<SafeModeEnvironment>>

pub fn read(&self) -> Result<Reader<SafeModeRoTransaction<'_>>>

Function used to obtain a “reader” which is used for read-only transactions.

pub fn write(&self) -> Result<Writer<'_>>

Function used to obtain a “writer” which is used for transactions. The writer.commit(); must be called to commit data added via the writer.