Crate nimbus

Re-exports

pub use error::NimbusError;

pub use error::Result;

pub use stateful::matcher::AppContext;

pub use evaluator::TargetingAttributes;

pub use schema::*;

pub use stateful::nimbus_client::*;

Modules

error

Not complete yet This is where the error definitions can go TODO: Implement proper error handling, this would include defining the error enum, impl std::error::Error using thiserror and ensuring all errors are handled appropriately

metrics

schema

stateful

Structs

EnrolledFeature

NimbusTargetingHelper

RemoteSettingsConfig

Custom configuration for the client. Currently includes the following:

Enums

EnrollmentStatus

RemoteSettingsServer

The Remote Settings server that the client should use.

Functions

evaluate_enrollment

Determine the enrolment status for an experiment.