Struct nimbus::NimbusClient

pub struct NimbusClient { /* fields omitted */ }

Nimbus is the main struct representing the experiments state It should hold all the information needed to communicate a specific user’s experimentation status

Implementations

impl NimbusClient

pub fn new<P: Into<PathBuf>>(
app_context: AppContext,
db_path: P,
config: Option<RemoteSettingsConfig>,
available_randomization_units: AvailableRandomizationUnits
) -> Result<Self>

pub fn initialize(&self) -> Result<()>

pub fn get_experiment_branch(&self, slug: String) -> Result<Option<String>>

pub fn get_feature_config_variables(
&self,
feature_id: String
) -> Result<Option<String>>

pub fn get_experiment_branches(
&self,
slug: String
) -> Result<Vec<ExperimentBranch>>

pub fn get_global_user_participation(&self) -> Result<bool>

pub fn set_global_user_participation(
&self,
user_participating: bool
) -> Result<Vec<EnrollmentChangeEvent>>

pub fn get_active_experiments(&self) -> Result<Vec<EnrolledExperiment>>

pub fn get_all_experiments(&self) -> Result<Vec<Experiment>>

pub fn get_available_experiments(&self) -> Result<Vec<AvailableExperiment>>

pub fn opt_in_with_branch(
&self,
experiment_slug: String,
branch: String
) -> Result<Vec<EnrollmentChangeEvent>>

pub fn opt_out(
&self,
experiment_slug: String
) -> Result<Vec<EnrollmentChangeEvent>>

pub fn update_experiments(&self) -> Result<Vec<EnrollmentChangeEvent>>

pub fn fetch_experiments(&self) -> Result<()>

pub fn apply_pending_experiments(&self) -> Result<Vec<EnrollmentChangeEvent>>

pub fn set_experiments_locally(&self, experiments_json: String) -> Result<()>

pub fn reset_telemetry_identifiers(
&self,
new_randomization_units: AvailableRandomizationUnits
) -> Result<Vec<EnrollmentChangeEvent>>

Reset internal state in response to application-level telemetry reset.

When the user resets their telemetry state in the consuming application, we need learn the new values of any external randomization units, and we need to reset any unique identifiers used internally by the SDK. If we don’t then we risk accidentally tracking across the telemetry reset, since we could use Nimbus metrics to link their pings from before and after the reset.

pub fn nimbus_id(&self) -> Result<Uuid>

pub fn set_nimbus_id(&self, uuid: &Uuid) -> Result<()>